pyproject.toml
[project]
name = "smart-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[a2a,eval]>=1.12.0",
    "google-cloud-aiplatform>=1.71.1",
    "notion-client",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "pydantic",
    "google-genai",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
